requirements.txt
PyYAML>=6.0
Pillow>=10.0.0  # 用于图片处理
python-dotenv>=1.0.0  # 用于环境变量支持

# 外部工具依赖说明：
# - Hugo静态站点生成器 (https://gohugo.io/)
# - Git (用于版本控制和部署)
# - SSH (用于连接GitHub仓库)
